+++ extension.ts
@@ -1,4 +1,5 @@
 import registerParleyExtension from "./index.ts";
 
 export { registerParleyExtension };
+export type { ParleyExtensionOptions, ParleyPresenceNameContext } from "./index.ts";
 export default registerParleyExtension;